Add Backend Error Trapping
So we don't get an error 500 on validation, error catching has been
added. We can do whatever we want with this on the frontend. Right
now the "confirm" box does not have functionality to make use of the
server response ({success: false, message: 'XXXX'}) but at least it's
there to use :)