Refactor the OTP code into classes.
A general OTP class was added, and the Yubikey support was refactored
to utilise it. This change also limited users to one valid OTP per
account in order to make for cleaner code (and because I don't know of
any other service that lets you have multiple different keys set up for
two factor authentication).

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>