Add Two Factor Authentication to the login process.
This changes the login process significantly; as we can no longer use
the authenticate function in brave.core.account.authenticate as the
sole function for verifying user login credentials. The username/email
and password code that was previously in authenticate has been moved
to verify_credentials, and all that remains in authenticate is the code
that updates the user's login history, duplicate detection, and session
clean up.

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>