Add support for for adding all characters to a grant.
The all characters change is significant because it will automatically
authorize characters to the grant as they're added. This is ideal for
applications which *must* have access to every character on an account,
such as an HR application or anti-spy tracking application. A bug that
prevented users from removing characters from a grant was also fixed.

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>