Remove sign in when authenticating already authorized apps.
When a user has an existing Core session, if they attempt to
authenticate an app they've already authorized, they will no longer be
prompted for their username and password. This reduces the amount that
users need to type their username and password while ensuring that they
still must do so before authorizing an application.

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>