Remove the binary seach from set_grants_permission.
The binary search algorithm didn't properly support wildcard
permissions with a wildcard in the middle, and the performance
improvement provided by the new algorithm was so negligible at tested
values that it wasn't worth resolving that issue. I'm leaving the
commits in the PR however, so that if we ever need to go back and fix
it due to permission issues we'll have it in the history.

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>