Add domain-wide cookie for storing user authorized apps.
Core now stores the id of every app authorized by the user in a cookie,
allowing applications to determine if they should show the user an
index page with the option to authenticate, or if the application
should just make the authorize call and then redirect the user to Core
to finish the authorization. This allows for seamless SSO features
without having to share sessions between applications, allowing for
untrusted applications to partake in the SSO environment.

Signed-off-by: Tyler O'Meara <Tyler@TylerOMeara.com>