Refactor the authorization mechanism to support multiple authorization types.
Adds a generic Authorization Method class that supplies an interface to
be implemented by subclasses to provide the necessary authorization
checks in various authorization methods. The authorization request flow
has been refactored to have the Core checks performed in a central
controller, and have the authorization method specific checks
implemented via hooks in the standard controller.